Done some major code cleanup to reduce repeated code. Implemented a removeGrammar method to fix a bug with outputting word probabilities and frequencies, removed forbidden characters in the same method.
Implemented the get word probability method. Changed the main menu to use a switch statement instead of an IF statement. Added a quit feature to the main menu.

Changed the comments above the methods to be summary comments instead of regular comments

Added small comments as notes to myself or to explain the operation of various complex parts of the code.