Added the loadpriortraining method to load previously saved training information.
Created an AIConsult class to write all of the formula's and such needed in the future for the actual consultation part of the software.

Sorted a major issue with the training not writing all of the information to the text files causing crashes when it came to loading that data back from the text file.