Made a link between training and consulting. Done the formula for sorting out consultation. Added the probability value to the party class.
Fixed a bug with reading in the file information. Noticed a MAJOR flaw with the way i load in data and the way data is consulted. This will need to be fixed by using serialization which will require the editing of the party class to allow it to be serialized.